+++ src/main/java/dk/kb/datahandler/kaltura/KalturaDeltaUploadJob.java
@@ -253,6 +253,7 @@ public class KalturaDeltaUploadJob {
      * 
      * @param  kalturaEntryId  The internal kaltura entryId
      * @throws IOException If API error
+     * @throws APIException
      */
     public static void deleteStream(String kalturaEntryId) throws IOException, APIException {
         initKalturaClient();            
@@ -272,7 +273,7 @@ public class KalturaDeltaUploadJob {
      * @param file_id Our reference to the stream.
      * @return kalturaId or null if does not exist.
      * @throws IOException
-     * 
+     * @throws APIException
      */
     private static String getInternalIdKaltura(String file_id) throws IOException, APIException {