+++ src/main/java/dk/kb/datahandler/oai/OaiHarvestClient.java
@@ -4,8 +4,6 @@ import java.io.IOException;
 import java.io.StringReader;
 import java.io.StringWriter;
 import java.io.Writer;
-import java.net.Authenticator;
-import java.net.PasswordAuthentication;
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
@@ -20,7 +18,6 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 import dk.kb.datahandler.config.ServiceConfig;
-import dk.kb.datahandler.job.JobCache;
 import dk.kb.util.webservice.exception.InternalServiceException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,7 +27,7 @@ import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
 
-import dk.kb.datahandler.model.v1.DsDatahandlerJobDto;
+import dk.kb.datahandler.model.v1.JobDto;
 import dk.kb.datahandler.model.v1.OaiTargetDto;
 import dk.kb.util.xml.XMLEscapeSanitiser;
 
@@ -41,14 +38,12 @@ public class OaiHarvestClient {
 
     private static final Logger log = LoggerFactory.getLogger(OaiHarvestClient.class);
 
-    private DsDatahandlerJobDto dsDatahandlerJob = null;
     private OaiTargetDto oaiTarget = null;
     private boolean completed=false;
     private String resumptionToken=null;
     private String from;
 
-    public OaiHarvestClient(DsDatahandlerJobDto job, OaiTargetDto oaiTarget,String from){
-        this.dsDatahandlerJob=job;
+    public OaiHarvestClient(OaiTargetDto oaiTarget, String from) {
         this.oaiTarget=oaiTarget;
         this.from=from;
     }
@@ -84,7 +79,6 @@ public class OaiHarvestClient {
         String errorMessage = getErrorMessage(document);
          if (errorMessage != null && errorMessage.trim().length() >1) {                       
             log.info("Error message from OAI server when harvesting set:"+set +" message:"+errorMessage);                    
-            dsDatahandlerJob.setCompletedTime(JobCache.formatSystemMillis(System.currentTimeMillis()));            
             oaiResponse.setError(true);
             return oaiResponse;// will have no records
          }
@@ -115,7 +109,7 @@ public class OaiHarvestClient {
     /* Will construct the uri for next http request. Resumption token will be set if not null.
      * Also special coding since  Cumulus/Cups API is not OAI-PMH compliant. 
      */
-    private String addQueryParamsToUri(String uri,String set, String resumptionToken, String metadataPrefix, String from) {
+    private String addQueryParamsToUri(String uri, String set, String resumptionToken, String metadataPrefix, String from) {
 
         //For unknown reason cumulus/cups oai API failes if metaData+set parameter is repeated with resumptionToken! (bug)
         if (resumptionToken==null && set != null) { //COPS fails if set is still used with resumptiontoken
@@ -136,7 +130,7 @@ public class OaiHarvestClient {
         return uri;
     }
 
-    public ArrayList<OaiRecord> extractRecordsFromXml( Document document ) {
+    public ArrayList<OaiRecord> extractRecordsFromXml(Document document) {
         NodeList nList = document.getElementsByTagName("record"); 
 
         ArrayList<OaiRecord> records= new ArrayList<OaiRecord>();