+++ src/main/java/dk/kb/datahandler/webservice/Application_v1.java
@@ -19,13 +19,11 @@ public class Application_v1 extends javax.ws.rs.core.Application {
         OpenApiResource.setConfig(ServiceConfig.getConfig());
 
         return new HashSet<>(Arrays.asList(
-                JacksonJsonProvider.class,
+                CustomJacksonJsonProvider.class,
                 DsDatahandlerApiServiceImpl.class,
                 ServiceApiServiceImpl.class,
                 ServiceExceptionMapper.class,
                 OpenApiResource.class
         ));
     }
-
-
 }