+++ src/test/java/dk/kb/datahandler/oai/OaiHarvestClientIntegrationTest.java
@@ -10,8 +10,6 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
-import dk.kb.datahandler.job.JobCache;
-import dk.kb.datahandler.model.v1.DsDatahandlerJobDto;
 import dk.kb.datahandler.model.v1.OaiTargetDto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -69,9 +67,7 @@ public class OaiHarvestClientIntegrationTest {
         String from ="2021-01-01";
          */
 
-        DsDatahandlerJobDto job = JobCache.createNewOaiJob(oaiTarget,from);        
-
-        OaiHarvestClient client = new OaiHarvestClient(job,oaiTarget,from);
+        OaiHarvestClient client = new OaiHarvestClient(oaiTarget, from);
         OaiResponse r1 = client.next();
         assertEquals(1000, r1.getRecords().size());
         assertNotNull(r1.getResumptionToken());
@@ -79,8 +75,6 @@ public class OaiHarvestClientIntegrationTest {
         //Fetch next 1000        
         OaiResponse r2= client.next();
         assertEquals(1000, r2.getRecords().size());
-    
-        JobCache.finishJob(job, 0, false);
     }
 
 
@@ -100,10 +94,9 @@ public class OaiHarvestClientIntegrationTest {
         oaiTarget.setDatasource(conf.getString("integration.oaiTargets[1].datasource"));
         oaiTarget.setFilter(OaiTargetDto.FilterEnum.PRESERVICA);
         oaiTarget.setDateStampFormat(OaiTargetDto.DateStampFormatEnum.DATETIME);
-        DsDatahandlerJobDto job = JobCache.createNewOaiJob(oaiTarget,null);
 
 
-        OaiHarvestClient client = new OaiHarvestClient(job,oaiTarget,null);
+        OaiHarvestClient client = new OaiHarvestClient(oaiTarget, null);
         OaiResponse r1 = client.next();
         assertEquals(200, r1.getRecords().size()); //there is over 200 now. 200 is batch size.
         assertNotNull(r1.getResumptionToken());
@@ -118,7 +111,6 @@ public class OaiHarvestClientIntegrationTest {
         OaiResponse r3= client.next();
         assertEquals(200, r3.getRecords().size());
         
-        JobCache.finishJob(job, 0, false);
     }
 
 }