+++ src/test/java/dk/kb/datahandler/storage/JobStorageForUnitTests.java
@@ -1,0 +1,18 @@
+package dk.kb.datahandler.storage;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public class JobStorageForUnitTests extends JobStorage {
+
+    public void clearTables() throws SQLException {
+        try(PreparedStatement stmt = connection.prepareStatement("DELETE FROM JOBS")) {
+            stmt.executeUpdate();
+            commit();
+        }
+    }
+
+    public JobStorageForUnitTests() throws SQLException {
+        super();
+    }
+}